abstract out createOptions
